+++ scripts/lichessTools.js
@@ -142,6 +142,7 @@
       WhiteSmilingFace: '\u263A',
       WhiteFrowningFace: '\u2639',
       BlackSquare: '\u25A0',
+      PalmBranch: '\u2E19',
 
       toEntity: function(s) {
         let result='';